+++ src/Console/Command/FixCommand.php
@@ -11,6 +11,7 @@ use Normalizator\Configuration\Configurator;
 use Normalizator\Finder\File;
 use Normalizator\Finder\Finder;
 use Normalizator\Normalizator;
+use Normalizator\Util\Glob;
 use Normalizator\Util\Logger;
 use Normalizator\Util\Timer;
 use Symfony\Component\Console\Attribute\AsCommand;
@@ -33,6 +34,7 @@ use function file_exists;
 use function implode;
 use function iterator_count;
 use function memory_get_peak_usage;
+use function preg_match;
 use function round;
 use function sprintf;
 
@@ -49,6 +51,7 @@ class FixCommand extends Command
         private Normalizator $normalizator,
         private Timer $timer,
         private Logger $logger,
+        private Glob $glob,
     ) {
         parent::__construct();
     }
@@ -107,7 +110,7 @@ class FixCommand extends Command
             new InputOption('leading-eol', 'l', InputOption::VALUE_NONE, 'Trim redundant leading newlines.'),
             new InputOption('middle-eol', 'm', InputOption::VALUE_OPTIONAL, 'Trim redundant middle empty newlines.', false),
             new InputOption('name', 'a', InputOption::VALUE_NONE, 'Fix file and directory names.'),
-            new InputOption('not', null, InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY, 'Skip given path(s)', []),
+            new InputOption('not', null, InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY, 'Skip given paths', []),
             new InputOption('permissions', 'u', InputOption::VALUE_NONE, 'Fix file and directory permissions.'),
             new InputOption('space-before-tab', 's', InputOption::VALUE_NONE, 'Clean spaces before tabs in the initial part of the line.'),
             new InputOption('trailing-whitespace', 'w', InputOption::VALUE_NONE, 'Trim trailing whitespace characters.'),
@@ -180,6 +183,7 @@ class FixCommand extends Command
 
         // Set excluded paths.
         $excludes = [];
+        $regexExcludes = [];
         foreach ($paths as $path) {
             foreach ($input->getOption('not') as $exclude) {
                 if (file_exists($path . '/' . $exclude)) {
@@ -187,6 +191,10 @@ class FixCommand extends Command
                 } elseif (file_exists($exclude)) {
                     $excludeInfo = new File($exclude);
                 } else {
+                    if ($this->glob->isGlob($exclude)) {
+                        $regexExcludes[] = $this->glob->convertToRegex($exclude);
+                    }
+
                     continue;
                 }
 
@@ -194,13 +202,15 @@ class FixCommand extends Command
             }
         }
         $excludes = array_unique($excludes);
+        $regexExcludes = array_unique($regexExcludes);
 
         $count = 0;
         foreach ($paths as $path) {
             $iterator = $this->normalize(
                 $path,
-                $excludes,
                 $output,
+                $excludes,
+                $regexExcludes,
             );
             $count += iterator_count($iterator);
         }
@@ -284,12 +294,19 @@ class FixCommand extends Command
      */
     private function normalize(
         string $path,
-        array $excludes,
         OutputInterface $output,
+        ?array $excludes,
+        ?array $regexExcludes,
     ): Iterator {
-        $iterator = $this->finder->getTree($path, static function (File $file) use ($excludes) {
+        $iterator = $this->finder->getTree($path, static function (File $file) use ($excludes, $regexExcludes) {
             foreach ($excludes as $exclude) {
                 if ($file->getRealPath() === $exclude) {
+                    return false;
+                }
+            }
+
+            foreach ($regexExcludes as $exclude) {
+                if (1 === preg_match($exclude, $file->getRealPath() ?: '')) {
                     return false;
                 }
             }